[CITE-127] - Addressed review comments - removed group not found exception from deletion, moved repositories to citesphere-model
