+++ citesphere/src/main/java/edu/asu/diging/citesphere/core/service/impl/GroupManager.java
@@ -6,10 +6,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import edu.asu.diging.citesphere.core.exceptions.GroupDoesNotExistException;
-import edu.asu.diging.citesphere.core.mongo.CustomCitationGroupRepository;
+import edu.asu.diging.citesphere.core.service.ICitationCollectionManager;
+import edu.asu.diging.citesphere.core.service.ICitationManager;
 import edu.asu.diging.citesphere.core.service.IGroupManager;
 import edu.asu.diging.citesphere.core.zotero.IZoteroManager;
 import edu.asu.diging.citesphere.data.bib.CitationGroupRepository;
+import edu.asu.diging.citesphere.data.bib.CustomCitationGroupRepository;
 import edu.asu.diging.citesphere.model.bib.ICitationGroup;
 import edu.asu.diging.citesphere.model.bib.impl.CitationGroup;
 import edu.asu.diging.citesphere.user.IUser;
@@ -25,6 +27,12 @@ public class GroupManager implements IGroupManager {
     
     @Autowired
     private CustomCitationGroupRepository customGroupRepo;
+    
+    @Autowired
+    private ICitationManager citationManager;
+    
+    @Autowired
+    private ICitationCollectionManager collectionManager;
 
     /*
      * (non-Javadoc)
@@ -50,10 +58,9 @@ public class GroupManager implements IGroupManager {
     }
     
     @Override
-    public void deleteLocalGroupCopy(String groupId) throws GroupDoesNotExistException {
-        Long deleteCount = customGroupRepo.deleteByGroupId(Integer.parseInt(groupId));
-        if (deleteCount == 0L) {
-            throw new GroupDoesNotExistException("There is no group with id " + groupId);
-        }
+    public void deleteLocalGroupCopy(String groupId) {
+        customGroupRepo.deleteByGroupId(Integer.parseInt(groupId));
+        collectionManager.deleteLocalGroupCollections(groupId);
+        citationManager.deleteLocalGroupCitations(groupId);
     }
 }
